memory_framework/requirements.txt
pyyaml>=6.0
flask>=2.0.0
